// book-socials-ui/src/app/services/models/page-response-book-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BookResponse } from '../models/book-response';
export interface PageResponseBookResponse {
  first?: boolean;
  hydraMember?: Array<BookResponse>;
  hydraTotal?: number;
  last?: boolean;
  page?: number;
  size?: number;
  totalPages?: number;
}